feat: SM-154 ChangingSpeech 컴포넌트 구현
