+++ src/components/domain/SearchBot/index.tsx
@@ -7,14 +7,12 @@ import {
   StyledContainer,
   StyledBot,
   StyledInputBackground,
-  StyledTextBackground,
-  StyledInput,
-  StyledSpeechBubble
+  StyledInput
 } from './styled';
 import useClickAway from '@hooks/useClickAway';
-import ChangingText from '@domain/ChangingText';
+import ChangingSpeech from '@domain/ChangingSpeech';
 
-const bartenderConversation = [
+const BARTENDER_CONVERSATION = [
   '찾으시는 칵테일이 있다면 저를 클릭 해주세요!',
   '저는 국제 바텐더 협회의 정회원이랍니다! 믿거나 말거나~',
   '당신의 눈동자에 건배!',
@@ -58,13 +56,7 @@ const SearchBot = (): ReactElement => {
           />
         </StyledInputBackground>
       </StyledBot>
-      {!showInput && (
-        <StyledTextBackground>
-          <StyledSpeechBubble>
-            <ChangingText textList={bartenderConversation} />
-          </StyledSpeechBubble>
-        </StyledTextBackground>
-      )}
+      {!showInput && <ChangingSpeech textList={BARTENDER_CONVERSATION} />}
     </StyledContainer>
   );
 };
